DT-906: Moved and extended spy entity usage rule
